+++ modules/external-api-index/src/main/resources/OSGI-INF/message-receiver-acl.xml
