+++ modules/external-api-index/src/main/resources/OSGI-INF/message-receiver-theme.xml
