+++ modules/external-api/pom.xml
@@ -99,6 +99,11 @@
       <version>${project.version}</version>
     </dependency>
     <dependency>
+      <groupId>org.opencastproject</groupId>
+      <artifactId>opencast-external-api-index</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
       <groupId>org.apache.commons</groupId>
       <artifactId>commons-lang3</artifactId>
     </dependency>
@@ -189,6 +194,7 @@
             <Bundle-SymbolicName>${project.artifactId}</Bundle-SymbolicName>
             <Build-Number>${buildNumber}</Build-Number>
             <Import-Package>
+              org.opencastproject.external.index,
               org.opencastproject.matterhorn.search.impl;version=${project.version},
               org.opencastproject.rest;version=${project.version},
               org.opencastproject.index.service.message,
@@ -216,8 +222,6 @@
               org.opencastproject.external.api,
               org.opencastproject.external.common,
               org.opencastproject.external.endpoint,
-              org.opencastproject.external.impl,
-              org.opencastproject.external.impl.index,
               org.opencastproject.external.userdirectory
             </Export-Package>
             <Service-Component>
@@ -227,19 +231,10 @@
               OSGI-INF/external-group-loader.xml,
               OSGI-INF/external-role-provider.xml,
               OSGI-INF/external-groups-endpoint.xml,
-              OSGI-INF/external-index.xml,
               OSGI-INF/external-security-endpoint.xml,
               OSGI-INF/external-series-endpoint.xml,
               OSGI-INF/external-workflow-definitions-endpoint.xml,
-              OSGI-INF/external-workflows-endpoint.xml,
-              OSGI-INF/message-receiver-acl.xml,
-              OSGI-INF/message-receiver-assetmanager.xml,
-              OSGI-INF/message-receiver-comment.xml,
-              OSGI-INF/message-receiver-group.xml,
-              OSGI-INF/message-receiver-scheduler.xml,
-              OSGI-INF/message-receiver-series.xml,
-              OSGI-INF/message-receiver-theme.xml,
-              OSGI-INF/message-receiver-workflow.xml
+              OSGI-INF/external-workflows-endpoint.xml
             </Service-Component>
           </instructions>
         </configuration>
