+++ modules/external-api/src/main/java/org/opencastproject/external/endpoint/BaseEndpoint.java
@@ -29,7 +29,7 @@ import static org.opencastproject.util.RestUtil.getEndpointUrl;
 
 import org.opencastproject.external.common.ApiMediaType;
 import org.opencastproject.external.common.ApiVersion;
-import org.opencastproject.external.impl.index.ExternalIndex;
+import org.opencastproject.external.index.ExternalIndex;
 import org.opencastproject.rest.RestConstants;
 import org.opencastproject.security.api.Organization;
 import org.opencastproject.security.api.Role;
