+++ modules/external-api/src/main/java/org/opencastproject/external/endpoint/GroupsEndpoint.java
@@ -39,7 +39,7 @@ import static org.opencastproject.util.doc.rest.RestParameter.Type.STRING;
 
 import org.opencastproject.external.common.ApiMediaType;
 import org.opencastproject.external.common.ApiResponses;
-import org.opencastproject.external.impl.index.ExternalIndex;
+import org.opencastproject.external.index.ExternalIndex;
 import org.opencastproject.index.service.api.IndexService;
 import org.opencastproject.index.service.impl.index.group.Group;
 import org.opencastproject.matterhorn.search.SearchIndexException;
