+++ modules/external-api/src/main/resources/OSGI-INF/external-base-endpoint.xml
@@ -16,7 +16,7 @@
              policy="static"
              bind="setSecurityService"/>
   <reference name="ExternalIndex"
-             interface="org.opencastproject.external.impl.index.ExternalIndex"
+             interface="org.opencastproject.external.index.ExternalIndex"
              cardinality="1..1"
              policy="static"
              bind="setExternalIndex"/>
