+++ modules/external-api/src/main/resources/OSGI-INF/external-events-endpoint.xml
@@ -23,7 +23,7 @@
              bind="addCatalogUIAdapter"
              unbind="removeCatalogUIAdapter"/>
   <reference name="ExternalIndex"
-             interface="org.opencastproject.external.impl.index.ExternalIndex"
+             interface="org.opencastproject.external.index.ExternalIndex"
              cardinality="1..1"
              policy="static"
              bind="setExternalIndex"/>
