+++ modules/external-api/src/main/resources/OSGI-INF/external-groups-endpoint.xml
@@ -11,7 +11,7 @@
     <provide interface="org.opencastproject.external.endpoint.GroupsEndpoint"/>
   </service>
   <reference name="ExternalIndex"
-             interface="org.opencastproject.external.impl.index.ExternalIndex"
+             interface="org.opencastproject.external.index.ExternalIndex"
              cardinality="1..1"
              policy="static"
              bind="setExternalIndex"/>
